Reworked code
